+++ R/arsenal.R
@@ -36,7 +36,6 @@
 NULL
 
 #### commands to build the package using devtools
-# devtools::document()
 # devtools::check_man()
 # devtools::test()
 # devtools::check()